rename: PotForm 이름 변경
