+++ src/pages/CreatePot/CreatePot.tsx
@@ -1,7 +1,7 @@
 import {
   mainContainer,
 } from "./CreatePot.style";
-import { EditPotForm } from "./components";
+import { PotForm } from "./components";
 import useCreatePot from "apis/hooks/pots/useCreatePot";
 
 const CreatePot = () => {
@@ -9,7 +9,7 @@ const CreatePot = () => {
 
   return (
     <main css={mainContainer}>
-      <EditPotForm type="create" onCompleted={(data) => mutate(data)} />
+      <PotForm type="create" onCompleted={(data) => mutate(data)} />
     </main>
   );
 };
